requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-SocketIO>=5.0
python-socketio>=5.0
paho-mqtt>=1.6
requests>=2.25